requirements.txt
torch
torchvision
facenet-pytorch
opencv-python
numpy
pillow
scikit-learn
kagglehub